Moved components to separate files
