+++ src/app/components/EnhancedTable/EnhancedTableHead.tsx
@@ -1,0 +1,57 @@
+import React from 'react';
+import {
+  TableHead, TableRow, TableCell, TableSortLabel, Box,
+} from '@mui/material';
+import { visuallyHidden } from '@mui/utils';
+import { HeadCell, Order } from './EnhancedTableTypes';
+
+interface EnhancedTableHeadProps {
+  onRequestSort: (event: React.MouseEvent<unknown>, property: keyof any) => void;
+  order: Order;
+  orderBy: string;
+  headCells: HeadCell[]
+}
+
+function EnhancedTableHead(props: EnhancedTableHeadProps) {
+  const {
+    order, orderBy, onRequestSort, headCells,
+  } = props;
+  // eslint-disable-next-line max-len
+  const createSortHandler = (property: keyof any) => (event: React.MouseEvent<unknown>) => {
+    onRequestSort(event, property);
+  };
+
+  return (
+    <TableHead>
+      <TableRow>
+        {headCells.map((headCell) => (
+          <TableCell
+            key={headCell.id as string}
+            align={headCell.align ? headCell.align : 'left'}
+            padding={headCell.disablePadding ? 'none' : 'normal'}
+            sortDirection={orderBy === headCell.id ? order : false}
+          >
+            {headCell.disableSorting
+              ? headCell.label
+              : (
+                <TableSortLabel
+                  active={orderBy === headCell.id}
+                  direction={orderBy === headCell.id ? order : 'asc'}
+                  onClick={createSortHandler(headCell.id)}
+                >
+                  {headCell.label}
+                  {orderBy === headCell.id ? (
+                    <Box component="span" sx={visuallyHidden}>
+                      {order === 'desc' ? 'sorted descending' : 'sorted ascending'}
+                    </Box>
+                  ) : null}
+                </TableSortLabel>
+              )}
+          </TableCell>
+        ))}
+      </TableRow>
+    </TableHead>
+  );
+}
+
+export default EnhancedTableHead;
