+++ src/app/components/JiraItemDetailsTable.tsx
@@ -6,8 +6,9 @@ import KeyboardArrowDownIcon from '@mui/icons-material/KeyboardArrowDown';
 import { useRecoilState, useRecoilValue } from 'recoil';
 import { JiraIssue } from '../../shared/dto/JiraTypes';
 import { jiraUpdatesState } from '../state/JiraState';
-import { EnhancedTable, HeadCell } from './EnhancedTable/EnhancedTable';
 import JiraItemDetails from './JiraItemDetails';
+import EnhancedTable from './EnhancedTable/EnhancedTable';
+import { HeadCell } from './EnhancedTable/EnhancedTableTypes';
 
 const KeyboardArrowUpIcon = styled(KeyboardArrowDownIcon)({
   transform: 'rotate(180deg)',
