+++ src/app/components/JiraItemsTable.tsx
@@ -2,8 +2,9 @@ import React from 'react';
 import { useRecoilValue } from 'recoil';
 import { jiraHistoryState } from '../state/JiraState';
 import { ProjectCard } from './ProjectStyledComponents';
-import { EnhancedTable, HeadCell } from './EnhancedTable/EnhancedTable';
 import JiraItems from './JiraItems';
+import EnhancedTable from './EnhancedTable/EnhancedTable';
+import { HeadCell } from './EnhancedTable/EnhancedTableTypes';
 
 type Props = {
   projectKey: string;
